Move IArrayBuilder to scala.collection.mutable
